Journal spec passing for postgres, not writing tags yet
