+++ core/src/main/resources/reference.conf
@@ -116,6 +116,9 @@ jdbc-journal {
     journal {
       tableName = "journal"
       schemaName = ""
+
+      # Only used in pre 5.0.0 Dao. Renaming individual column names is
+      # no longer supported
       columnNames {
         ordering = "ordering"
         deleted = "deleted"
@@ -133,7 +136,9 @@ jdbc-journal {
   # in order to keep consistent behavior over write/read sides
   tagSeparator = ${akka-persistence-jdbc.tagSeparator}
 
-  dao = "akka.persistence.jdbc.journal.dao.ByteArrayJournalDao"
+  # If you have data from pre 5.0.0 use the legacy akka.persistence.jdbc.journal.dao.legacy.ByteArrayJournalDao
+  # Dao. Migration to the new dao will be added in the future.
+  dao = "akka.persistence.jdbc.journal.dao.AkkaSerializerJournalDao"
 
   # The size of the buffer used when queueing up events for batch writing. This number must be bigger then the number
   # of events that may be written concurrently. In other words this number must be bigger than the number of persistent
