+++ core/src/main/scala/akka/persistence/jdbc/journal/dao/JournalTables.scala
@@ -1,13 +1,27 @@
-/*
- * Copyright (C) 2014 - 2019 Dennis Vriend <https://github.com/dnvriend>
- * Copyright (C) 2019 - 2020 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package akka.persistence.jdbc
-package journal.dao
+package akka.persistence.jdbc.journal.dao
 
 import akka.persistence.jdbc.config.JournalTableConfiguration
+import akka.persistence.jdbc.journal.dao.JournalTables.JournalAkkaSerializationRow
 
+object JournalTables {
+  case class JournalAkkaSerializationRow(
+      ordering: Long,
+      deleted: Boolean, // TODO, remove? Or do we need it for something
+      persistenceId: String,
+      sequenceNumber: Long,
+      writer: String,
+      eventManifest: String,
+      eventPayload: Array[Byte],
+      eventSerId: Int,
+      eventSerManifest: String,
+      metaPayload: Option[Array[Byte]],
+      metaSerId: Option[Int],
+      metaSerManifest: Option[String])
+}
+
+/**
+ * For the schema dded in 5.0.0
+ */
 trait JournalTables {
   val profile: slick.jdbc.JdbcProfile
 
@@ -16,20 +30,42 @@ trait JournalTables {
   def journalTableCfg: JournalTableConfiguration
 
   class Journal(_tableTag: Tag)
-      extends Table[JournalRow](
+      extends Table[JournalAkkaSerializationRow](
         _tableTag,
         _schemaName = journalTableCfg.schemaName,
         _tableName = journalTableCfg.tableName) {
-    def * = (ordering, deleted, persistenceId, sequenceNumber, message, tags) <> (JournalRow.tupled, JournalRow.unapply)
+    def * =
+      (
+        ordering,
+        deleted,
+        persistenceId,
+        sequenceNumber,
+        writer,
+        eventManifest,
+        eventPayload,
+        eventSerId,
+        eventSerManifest,
+        metaPayload,
+        metaSerId,
+        metaSerManifest) <> (JournalAkkaSerializationRow.tupled, JournalAkkaSerializationRow.unapply)
 
-    val ordering: Rep[Long] = column[Long](journalTableCfg.columnNames.ordering, O.AutoInc)
+    val ordering: Rep[Long] = column[Long]("ordering", O.AutoInc)
     val persistenceId: Rep[String] =
-      column[String](journalTableCfg.columnNames.persistenceId, O.Length(255, varying = true))
-    val sequenceNumber: Rep[Long] = column[Long](journalTableCfg.columnNames.sequenceNumber)
-    val deleted: Rep[Boolean] = column[Boolean](journalTableCfg.columnNames.deleted, O.Default(false))
-    val tags: Rep[Option[String]] =
-      column[Option[String]](journalTableCfg.columnNames.tags, O.Length(255, varying = true))
-    val message: Rep[Array[Byte]] = column[Array[Byte]](journalTableCfg.columnNames.message)
+      column[String]("persistence_id", O.Length(255, varying = true))
+    val sequenceNumber: Rep[Long] = column[Long]("sequence_number")
+    val deleted: Rep[Boolean] = column[Boolean]("deleted", O.Default(false))
+
+    val writer: Rep[String] = column[String]("writer")
+    val eventManifest: Rep[String] = column[String]("event_manifest")
+
+    val eventPayload: Rep[Array[Byte]] = column[Array[Byte]]("event_payload")
+    val eventSerId: Rep[Int] = column[Int]("event_ser_id")
+    val eventSerManifest: Rep[String] = column[String]("event_ser_manifest")
+
+    val metaPayload: Rep[Option[Array[Byte]]] = column[Option[Array[Byte]]]("meta_payload")
+    val metaSerId: Rep[Option[Int]] = column[Option[Int]]("meta_ser_id")
+    val metaSerManifest: Rep[Option[String]] = column[Option[String]]("meta_ser_manifest")
+
     val pk = primaryKey(s"${tableName}_pk", (persistenceId, sequenceNumber))
     val orderingIdx = index(s"${tableName}_ordering_idx", ordering, unique = true)
   }
