+++ core/src/main/scala/akka/persistence/jdbc/package.scala
@@ -5,12 +5,4 @@
 
 package akka.persistence
 
-package object jdbc {
-  final case class JournalRow(
-      ordering: Long,
-      deleted: Boolean,
-      persistenceId: String,
-      sequenceNumber: Long,
-      message: Array[Byte],
-      tags: Option[String] = None)
-}
+package object jdb {}
