+++ core/src/test/resources/schema/mysql/mysql-schema-v5.sql
@@ -1,0 +1,23 @@
+DROP TABLE IF EXISTS journal;
+
+CREATE TABLE IF NOT EXISTS journal (
+  ordering SERIAL,
+  persistence_id VARCHAR(255) NOT NULL,
+  sequence_number BIGINT NOT NULL,
+  deleted BOOLEAN DEFAULT FALSE NOT NULL,
+  tags VARCHAR(255) DEFAULT NULL,
+  message BLOB NOT NULL,
+  PRIMARY KEY(persistence_id, sequence_number)
+);
+
+CREATE UNIQUE INDEX journal_ordering_idx ON journal(ordering);
+
+DROP TABLE IF EXISTS snapshot;
+
+CREATE TABLE IF NOT EXISTS snapshot (
+  persistence_id VARCHAR(255) NOT NULL,
+  sequence_number BIGINT NOT NULL,
+  created BIGINT NOT NULL,
+  snapshot BLOB NOT NULL,
+  PRIMARY KEY (persistence_id, sequence_number)
+);
