+++ core/src/test/scala/akka/persistence/jdbc/journal/dao/TagsSerializationTest.scala
@@ -3,7 +3,7 @@
  * Copyright (C) 2019 - 2020 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.persistence.jdbc.journal.dao
+package akka.persistence.jdbc.journal.dao.legacy
 
 import akka.persistence.jdbc.SharedActorSystemTestSpec
