+++ project/Dependencies.scala
@@ -9,8 +9,8 @@ object Dependencies {
   val Scala213 = "2.13.1"
   val ScalaVersions = Seq(Scala212, Scala213)
 
-  val AkkaVersion = "2.6.5"
-  val AkkaBinaryVersion = "2.6"
+  val AkkaVersion = "2.6.10"
+  val AkkaBinaryVersion = AkkaVersion.take(3)
 
   val SlickVersion = "3.3.3"
   val ScalaTestVersion = "3.1.4"
